Re-arranged the shared folder contents to make file structure more clear. Renamed typedef file
