+++ shared/lib/zotmeal.types.ts
